requirements.txt
beautifulsoup4>=4.13.4
httpx>=0.28.1
python-dotenv>=1.1.0